fix(klarna-sdk): added test cases for PrimerHeadlessKlarnaComponentTests
