+++ Sources/PrimerSDK/Classes/Core/PrimerHeadlessUniversalCheckout/Composable/Klarna/Components/KlarnaTokenizationComponent.swift
@@ -18,25 +18,38 @@ protocol KlarnaTokenizationComponentProtocol: KlarnaTokenizationManagerProtocol 
 
 class KlarnaTokenizationComponent: KlarnaTokenizationManager, KlarnaTokenizationComponentProtocol {
     // MARK: - Properties
+
     private let paymentMethod: PrimerPaymentMethod
     private let apiClient: PrimerAPIClientProtocol
     private var clientSession: ClientSession.APIResponse?
     private var paymentSessionId: String?
     private var recurringPaymentDescription: String?
+
     // MARK: - Settings
+
     let settings: PrimerSettingsProtocol = DependencyContainer.resolve()
+
     // MARK: - Init
-    init(paymentMethod: PrimerPaymentMethod) {
+
+    init(
+        paymentMethod: PrimerPaymentMethod,
+        tokenizationService: TokenizationServiceProtocol = TokenizationService()
+    ) {
         self.paymentMethod = paymentMethod
-        self.apiClient = PrimerAPIConfigurationModule.apiClient ?? PrimerAPIClient()
-        self.clientSession = PrimerAPIConfigurationModule.apiConfiguration?.clientSession
+        apiClient = PrimerAPIConfigurationModule.apiClient ?? PrimerAPIClient()
+        clientSession = PrimerAPIConfigurationModule.apiConfiguration?.clientSession
         let klarnaOptions = PrimerSettings.current.paymentMethodOptions.klarnaOptions
-        self.recurringPaymentDescription = klarnaOptions?.recurringPaymentDescription
-        super.init(createResumePaymentService: CreateResumePaymentService(paymentMethodType: paymentMethod.type))
+        recurringPaymentDescription = klarnaOptions?.recurringPaymentDescription
+
+        super.init(
+            tokenizationService: tokenizationService,
+            createResumePaymentService: CreateResumePaymentService(paymentMethodType: paymentMethod.type),
+        )
     }
 }
 
 // MARK: - Validate
+
 extension KlarnaTokenizationComponent {
     func validate() throws {
         guard
@@ -59,6 +72,7 @@ extension KlarnaTokenizationComponent {
             break
         }
     }
+
     /// - Validates the necessary conditions specific to one-off payment operations.
     func validateOneOffPayment() throws {
         if AppState.current.amount == nil {
@@ -71,13 +85,14 @@ extension KlarnaTokenizationComponent {
         if lineItems.isEmpty {
             throw KlarnaHelpers.getInvalidValueError(key: "lineItems")
         }
-        if !(lineItems.filter({ $0.amount == nil })).isEmpty {
+        if !(lineItems.filter { $0.amount == nil }).isEmpty {
             throw KlarnaHelpers.getInvalidValueError(key: "settings.orderItems")
         }
     }
 }
 
 // MARK: - Create payment session
+
 extension KlarnaTokenizationComponent {
     func createPaymentSession() -> Promise<Response.Body.Klarna.PaymentSession> {
         return Promise { seal in
@@ -96,15 +111,15 @@ extension KlarnaTokenizationComponent {
             firstly {
                 requestPrimerConfiguration(decodedJWTToken: decodedJWTToken, request: requestUpdateBody)
             }
-            .then({ () -> Promise<Response.Body.Klarna.PaymentSession> in
+            .then { () -> Promise<Response.Body.Klarna.PaymentSession> in
                 // Prepare the body for the payment session creation request
                 let body = self.prepareKlarnaPaymentSessionRequestBody(paymentMethodConfigId: paymentMethodConfigId)
                 // Create the Klarna payment session
                 return self.createKlarnaSession(with: body, decodedJWTToken: decodedJWTToken)
-            })
-            .done({ paymentSession in
+            }
+            .done { paymentSession in
                 seal.fulfill(paymentSession)
-            })
+            }
             .catch { error in
                 seal.reject(error)
             }
@@ -113,6 +128,7 @@ extension KlarnaTokenizationComponent {
 }
 
 // MARK: - Authorize payment session
+
 extension KlarnaTokenizationComponent {
     func authorizePaymentSession(authorizationToken: String) -> Promise<Response.Body.Klarna.CustomerToken> {
         return Promise { seal in
@@ -162,6 +178,7 @@ extension KlarnaTokenizationComponent {
 }
 
 // MARK: - Klarna Creation helpers
+
 private extension KlarnaTokenizationComponent {
     /// - Helper method to prepare Klarna Payment Session request body
     private func prepareKlarnaPaymentSessionRequestBody(paymentMethodConfigId: String) -> Request.Body.Klarna.CreatePaymentSession {
@@ -170,14 +187,17 @@ private extension KlarnaTokenizationComponent {
             with: paymentMethodConfigId,
             clientSession: clientSession,
             recurringPaymentDescription: recurringPaymentDescription,
-            redirectUrl: urlScheme)
+            redirectUrl: urlScheme
+        )
     }
+
     /// - Helper method to prepare Client Session Update Request body with actions
     private func prepareKlarnaClientSessionActionsRequestBody() -> ClientSessionUpdateRequest {
         let params: [String: Any] = ["paymentMethodType": paymentMethod.type]
         let actions = [ClientSession.Action.selectPaymentMethodActionWithParameters(params)]
         return ClientSessionUpdateRequest(actions: ClientSessionAction(actions: actions))
     }
+
     /// - Request to update Primer Configuration with actions
     /// - Sets the client session with updated primer configuration request data
     private func requestPrimerConfiguration(decodedJWTToken: DecodedJWTToken, request: ClientSessionUpdateRequest) -> Promise<Void> {
@@ -196,9 +216,11 @@ private extension KlarnaTokenizationComponent {
             }
         }
     }
+
     /// - Request to create  Klarna Payment Session
     /// - Sets the 'paymentSessionId'  with response's 'sessionId'
-    private func createKlarnaSession(with body: Request.Body.Klarna.CreatePaymentSession, decodedJWTToken: DecodedJWTToken) -> Promise<Response.Body.Klarna.PaymentSession> {
+    private func createKlarnaSession(with body: Request.Body.Klarna.CreatePaymentSession,
+                                     decodedJWTToken: DecodedJWTToken) -> Promise<Response.Body.Klarna.PaymentSession> {
         return Promise { seal in
             apiClient.createKlarnaPaymentSession(clientToken: decodedJWTToken,
                                                  klarnaCreatePaymentSessionAPIRequest: body) { [weak self] result in
@@ -216,18 +238,23 @@ private extension KlarnaTokenizationComponent {
 }
 
 // MARK: - Klarna Authorization helpers
+
 extension KlarnaTokenizationComponent {
     /// - Helper method to prepare Klarna Finalize Payment Session body
-    private func prepareKlarnaFinalizePaymentSessionBody(paymentMethodConfigId: String, sessionId: String) -> Request.Body.Klarna.FinalizePaymentSession {
+    private func prepareKlarnaFinalizePaymentSessionBody(paymentMethodConfigId: String, sessionId: String) -> Request.Body.Klarna
+        .FinalizePaymentSession {
         return KlarnaHelpers.getKlarnaFinalizePaymentBody(
             with: paymentMethodConfigId,
-            sessionId: sessionId)
+            sessionId: sessionId
+        )
     }
+
     /// - Request to finalize  Klarna Payment Session
-    private func finalizeKlarnaPaymentSession(with clientToken: DecodedJWTToken, body: Request.Body.Klarna.FinalizePaymentSession) -> Promise<Response.Body.Klarna.CustomerToken> {
+    private func finalizeKlarnaPaymentSession(with clientToken: DecodedJWTToken,
+                                              body: Request.Body.Klarna.FinalizePaymentSession) -> Promise<Response.Body.Klarna.CustomerToken> {
         return Promise { seal in
             apiClient.finalizeKlarnaPaymentSession(clientToken: clientToken,
-                                                   klarnaFinalizePaymentSessionRequest: body) { (result) in
+                                                   klarnaFinalizePaymentSessionRequest: body) { result in
                 switch result {
                 case .success(let response):
                     seal.fulfill(response)
@@ -237,19 +264,24 @@ extension KlarnaTokenizationComponent {
             }
         }
     }
+
     /// - Helper method to prepare Klarna Customer Token body
-    private func prepareKlarnaCustomerTokenBody(paymentMethodConfigId: String, sessionId: String, authorizationToken: String) -> Request.Body.Klarna.CreateCustomerToken {
+    private func prepareKlarnaCustomerTokenBody(paymentMethodConfigId: String, sessionId: String, authorizationToken: String) -> Request.Body.Klarna
+        .CreateCustomerToken {
         return KlarnaHelpers.getKlarnaCustomerTokenBody(
             with: paymentMethodConfigId,
             sessionId: sessionId,
             authorizationToken: authorizationToken,
-            recurringPaymentDescription: recurringPaymentDescription)
+            recurringPaymentDescription: recurringPaymentDescription
+        )
     }
+
     /// - Request to create  Klarna Customer Token
-    private func createKlarnaCustomerToken(with clientToken: DecodedJWTToken, body: Request.Body.Klarna.CreateCustomerToken) -> Promise<Response.Body.Klarna.CustomerToken> {
+    private func createKlarnaCustomerToken(with clientToken: DecodedJWTToken,
+                                           body: Request.Body.Klarna.CreateCustomerToken) -> Promise<Response.Body.Klarna.CustomerToken> {
         return Promise { seal in
             apiClient.createKlarnaCustomerToken(clientToken: clientToken,
-                                                klarnaCreateCustomerTokenAPIRequest: body) { (result) in
+                                                klarnaCreateCustomerTokenAPIRequest: body) { result in
                 switch result {
                 case .success(let response):
                     seal.fulfill(response)
@@ -262,6 +294,7 @@ extension KlarnaTokenizationComponent {
 }
 
 // MARK: - Klarna Testing helpers
+
 extension KlarnaTokenizationComponent {
     func setSessionId(paymentSessionId: String) {
         self.paymentSessionId = paymentSessionId
