+++ Sources/PrimerSDK/Classes/Core/PrimerHeadlessUniversalCheckout/Composable/Klarna/Managers/KlarnaTokenizationManager.swift
@@ -18,11 +18,11 @@ protocol KlarnaTokenizationManagerProtocol {
      object on successful tokenization or rejects with an `Error` if the tokenization process fails.
      */
     func tokenizeHeadless(customerToken: Response.Body.Klarna.CustomerToken?, offSessionAuthorizationId: String?) -> Promise<PrimerCheckoutData>
-    func tokenizeDropIn(customerToken: Response.Body.Klarna.CustomerToken?, offSessionAuthorizationId: String?) -> Promise<PrimerPaymentMethodTokenData>
+    func tokenizeDropIn(customerToken: Response.Body.Klarna.CustomerToken?, offSessionAuthorizationId: String?)
+        -> Promise<PrimerPaymentMethodTokenData>
 }
 
 class KlarnaTokenizationManager: KlarnaTokenizationManagerProtocol {
-
     // MARK: - Properties
 
     private let tokenizationService: TokenizationServiceProtocol
@@ -30,13 +30,17 @@ class KlarnaTokenizationManager: KlarnaTokenizationManagerProtocol {
     private let createResumePaymentService: CreateResumePaymentServiceProtocol
 
     // MARK: - Init
-    init(tokenizationService: TokenizationServiceProtocol = TokenizationService(),
-         createResumePaymentService: CreateResumePaymentServiceProtocol) {
+
+    init(
+        tokenizationService: TokenizationServiceProtocol,
+        createResumePaymentService: CreateResumePaymentServiceProtocol
+    ) {
         self.tokenizationService = tokenizationService
         self.createResumePaymentService = createResumePaymentService
     }
 
     // MARK: - Tokenize Headless
+
     func tokenizeHeadless(customerToken: Response.Body.Klarna.CustomerToken?, offSessionAuthorizationId: String?) -> Promise<PrimerCheckoutData> {
         return Promise { seal in
             PrimerDelegateProxy.primerHeadlessUniversalCheckoutDidStartTokenization(for: PrimerPaymentMethodType.klarna.rawValue)
@@ -60,7 +64,9 @@ class KlarnaTokenizationManager: KlarnaTokenizationManagerProtocol {
     }
 
     // MARK: - Tokenize DropIn
-    func tokenizeDropIn(customerToken: Response.Body.Klarna.CustomerToken?, offSessionAuthorizationId: String?) -> Promise<PrimerPaymentMethodTokenData> {
+
+    func tokenizeDropIn(customerToken: Response.Body.Klarna.CustomerToken?,
+                        offSessionAuthorizationId: String?) -> Promise<PrimerPaymentMethodTokenData> {
         return Promise { seal in
             firstly {
                 getRequestBody(customerToken: customerToken, offSessionAuthorizationId: offSessionAuthorizationId)
@@ -105,7 +111,7 @@ extension KlarnaTokenizationManager {
                 firstly {
                     self.createPaymentEvent(token)
                 }
-                .done { paymentResponse -> Void in
+                .done { paymentResponse in
                     let paymentCheckoutData = PrimerCheckoutData(payment: PrimerCheckoutDataPayment(from: paymentResponse))
                     seal.fulfill(paymentCheckoutData)
                 }
@@ -122,7 +128,8 @@ extension KlarnaTokenizationManager {
         return createResumePaymentService.createPayment(paymentRequest: paymentRequest)
     }
 
-    private func getRequestBody(customerToken: Response.Body.Klarna.CustomerToken?, offSessionAuthorizationId: String?) -> Promise<Request.Body.Tokenization> {
+    private func getRequestBody(customerToken: Response.Body.Klarna.CustomerToken?,
+                                offSessionAuthorizationId: String?) -> Promise<Request.Body.Tokenization> {
         return Promise { seal in
             var customerTokenId: String?
             var paymentInstrument: TokenizationRequestBodyPaymentInstrument
